Tests/LibWeb/Text/expected/URL/url.txt

ftp://serenityos.org:21
protocol => 'ftp:'
username => ''
password => ''
host => 'serenityos.org'
hostname => 'serenityos.org'
port => ''
pathname => '/'
search => ''
